# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Transaction:
    properties:
      ai_category_suggestion:
        type: string
      ai_confidence_score:
        type: number
      amount:
        type: number
      amount_in_base:
        type: number
      category_id:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      merchant:
        type: string
      metadata:
        type: string
      source:
        type: string
      transaction_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      currency_preference:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  dto.AIInsightResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      insight:
        type: string
      insight_type:
        type: string
      recommendation:
        type: string
      warning:
        type: string
    type: object
  dto.AnalysisSnapshotResponse:
    properties:
      category_breakdown:
        additionalProperties:
          type: number
        type: object
      created_at:
        type: string
      id:
        type: integer
      net_cashflow:
        type: number
      period_end:
        type: string
      period_start:
        type: string
      period_type:
        type: string
      top_expense_category:
        type: string
      total_expense:
        type: number
      total_income:
        type: number
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API for Spendly Personal Finance Tracker.
  title: Spendly API
  version: "1.0"
paths:
  /analysis/latest:
    get:
      description: Fetch the latest monthly analysis for a given user UUID
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AnalysisSnapshotResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get latest analysis snapshot
      tags:
      - analysis
  /insights/{id}/read:
    patch:
      description: Update insight is_read status to true
      parameters:
      - description: Insight ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Mark insight as read
      tags:
      - insights
  /insights/latest:
    get:
      description: Fetch recent AI-generated insights for a given user UUID
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      - description: Limit results (default 5)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AIInsightResponse'
            type: array
      summary: List AI Insights
      tags:
      - insights
  /transactions:
    get:
      description: Get a list of transactions for a specific user ID with pagination
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      - description: Results per page
        in: query
        name: limit
        type: integer
      - description: Offset results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Transaction'
            type: array
      summary: List Transactions
      tags:
      - transactions
  /users/{id}:
    get:
      description: Fetch profile details for a given User UUID
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get User Profile
      tags:
      - users
swagger: "2.0"